refactor: integrate dto
