--- src/articles/dto/update-article.dto.ts
@@ -1,5 +1,0 @@
-import { PartialType } from "@nestjs/swagger";
-
-import { CreateArticleDto } from "./create-article.dto";
-
-export class UpdateArticleDto extends PartialType(CreateArticleDto) {}
